requirements.txt
cachetools==3.1.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.3
oauth2client==4.1.3
oauthlib==3.0.1
requests==2.21.0
requests-oauthlib==1.2.0
